Separate packages and dataclass for ovdc
