+++ container_service_extension/def_/models.py
@@ -2,11 +2,9 @@
 # Copyright (c) 2017 VMware, Inc. All Rights Reserved.
 # SPDX-License-Identifier: BSD-2-Clause
 from dataclasses import dataclass
-from typing import List
 
 import container_service_extension.def_.utils as def_utils
 import container_service_extension.utils as utils
-import container_service_extension.exceptions as cse_exception
 
 
 @dataclass(frozen=True)
@@ -285,4 +283,3 @@ class DefEntity:
         self.entityType = entityType
         self.externalId = externalId
         self.state = state
-
